develops campsite model to include an arrayt of facilities, not yet populated
