+++ api/models/facility_models.py
@@ -1,19 +1,25 @@
-from database.database import Base
-from sqlalchemy import Column, Integer, String, ForeignKey
+# from database.database import Base
+# from sqlalchemy import Column, Integer, String, ForeignKey
+# from sqlalchemy.orm import relationship, Mapped
 
-class Facility(Base):
-    __tablename__ = "facilities"
+# from api.models.campsite_models import Campsite
 
-    facility_id = Column(Integer, primary_key=True)
-    facility_name = Column(String)
-    facility_img_url = Column(String)
+# class Facility(Base):
+#     __tablename__ = "facilities"
 
-class CampsiteFacility(Base):
-    __tablename__ = "campsite_facilities"
+#     facility_id = Column(Integer, primary_key=True)
+#     facility_name = Column(String)
+#     facility_img_url = Column(String)
 
-    campsite_facility_id = Column(Integer, primary_key=True)
-    campsite_id = Column(Integer, ForeignKey("campsites.campsite_id"))
-    facility_id = Column(Integer, ForeignKey("facilities.facility_id"))
+    # campsite: Mapped["Campsite"] = relationship("Campsite", back_populates="facilities")
+    
+
+# class CampsiteFacility(Base):
+#     __tablename__ = "campsite_facilities"
+
+#     campsite_facility_id = Column(Integer, primary_key=True)
+#     campsite_id = Column(Integer, ForeignKey("campsites.campsite_id"))
+#     facility_id = Column(Integer, ForeignKey("facilities.facility_id"))
